// frontend/src/api/model/memoEntryRead.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Qualis API
 * OpenAPI spec version: 0.1.0
 */
import type { MemoEntryReadParentType } from './memoEntryReadParentType';
import type { MemoEntryReadCreatedBy } from './memoEntryReadCreatedBy';
import type { MemoEntryReadLastEditedBy } from './memoEntryReadLastEditedBy';
import type { MemoCommentRead } from './memoCommentRead';

export interface MemoEntryRead {
    id: number;
    parent_type: MemoEntryReadParentType;
    parent_id: number;
    title: string;
    body: string;
    position: number;
    created_at: string;
    updated_at: string;
    created_by: MemoEntryReadCreatedBy;
    last_edited_by: MemoEntryReadLastEditedBy;
    comments: MemoCommentRead[];
}